# AVX2 optimization for Jpeg YCbCr color conversion