--- a/src/ImageSharp/Formats/Jpeg/Common/Decoder/ColorConverters/JpegColorConverter.FromGrayScale.cs
+++ b/src/ImageSharp/Formats/Jpeg/Common/Decoder/ColorConverters/JpegColorConverter.FromGrayScale.cs
@@ -1,23 +1,21 @@
 ﻿using System;
 using System.Numerics;
 
-namespace SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder
+namespace SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder.ColorConverters
 {
     internal abstract partial class JpegColorConverter
     {
-        internal class FromRgb : JpegColorConverter
+        internal class FromGrayScale : ColorConverters.JpegColorConverter
         {
-            public FromRgb()
-                : base(JpegColorSpace.RGB)
+            public FromGrayScale()
+                : base(JpegColorSpace.GrayScale)
             {
             }
 
             public override void ConvertToRGBA(ComponentValues values, Span<Vector4> result)
             {
                 // TODO: We can optimize a lot here with Vector<float> and SRCS.Unsafe()!
-                ReadOnlySpan<float> rVals = values.Component0;
-                ReadOnlySpan<float> gVals = values.Component1;
-                ReadOnlySpan<float> bVals = values.Component2;
+                ReadOnlySpan<float> yVals = values.Component0;
 
                 var v = new Vector4(0, 0, 0, 1);
 
@@ -25,13 +23,11 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder
 
                 for (int i = 0; i < result.Length; i++)
                 {
-                    float r = rVals[i];
-                    float g = gVals[i];
-                    float b = bVals[i];
+                    float y = yVals[i];
 
-                    v.X = r;
-                    v.Y = g;
-                    v.Z = b;
+                    v.X = y;
+                    v.Y = y;
+                    v.Z = y;
 
                     v *= scale;