… tests/ImageSharp.Benchmarks/Image/EncodeBmpMultiple.cs
… tests/ImageSharp.Benchmarks/Image/Jpeg/DecodeJpeg.cs
… tests/ImageSharp.Benchmarks/Image/Jpeg/DecodeJpegMultiple.cs
… tests/ImageSharp.Benchmarks/Image/Jpeg/EncodeJpeg.cs
… tests/ImageSharp.Benchmarks/Image/Jpeg/EncodeJpegMultiple.cs
… tests/ImageSharp.Benchmarks/Image/Jpeg/YCbCrColorConversion.cs
… tests/ImageSharp.Tests/Formats/Jpg/JpegColorConverterTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/JpegProfilingBenchmarks.cs