### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
- Managed to fight `Vector<T>` and now we have a special `JpegColorConverter.FromYCbCrSimdAvx2` execution path for AVX2-enabled (Haswell+) CPU-s with `Vector<float>.Count == 8`.
- The "old" `JpegColorConverter.FromYCbCrSimd` is capable of handling the `Vector<float>.Count == 4` path even when `Vector.IsHardwareAccelerated == false`. It's covered with it's own unit test.
- All converters were moved to the `ColorConverters` subfolder.

### Benchmarks

`JpegColorConverter.FromYCbCrSimdAvx2` is much faster on CPU-s supporting it:
```
      Method |       Mean |     Error |    StdDev | Scaled | Allocated |
------------ |-----------:|----------:|----------:|-------:|----------:|
      Scalar | 9,396.9 ns | 696.05 ns | 39.328 ns |   1.00 |       0 B |
 SimdVector4 | 1,687.6 ns |  77.52 ns |  4.380 ns |   0.18 |       0 B |
    SimdAvx2 |   410.9 ns | 116.94 ns |  6.608 ns |   0.04 |       0 B |
```

The [total decoder performance improvement](https://gist.github.com/antonfirsov/7e3545ed77fa863a58a75d255483e086#gistcomment-2208993) is also significant. The postprocessing step takes now less than `50%` of the full decoder CPU load.