@JimBobSquarePants 2017-09-21
TIL. `8388608` is the number of bits in a megabyte. How the number works in `FastRound` I still have no idea.