@antonfirsov 2017-09-21
@JimBobSquarePants some good points to start:
- [An online toy](https://www.h-schmidt.net/FloatConverter/IEEE754.html) to with the bits of an IEEE-754 single precision float
- [The matching Wikipedia article](https://en.wikipedia.org/wiki/Single-precision_floating-point_format)

TBH I haven't took the time to understand `FastRound` yet. For me, the only way is to take a pen, a paper + at least a half free hour to play with the bits.